Replace SSR scholarship routes with /api/v1/scholarship JSON API
The student portal now lives on Cloudflare Pages, so the Vapor server
no longer renders its pages. Everything that used to be a host-gated
HTML controller is now a JSON / form-redirect endpoint mounted under
/api/v1/scholarship and consumed by the static site's scholarship.js.

- ScholarshipAPIController consolidates the public, applicant, and
  organizer routes into a single RouteCollection. JSON for reads
  (`/info`, `/me`, `/me/application`, `/organizer/applications`,
  `/organizer/applications/:id`, `/organizer/budget`); 303 redirects
  back to STUDENT_BASE_URL for form POSTs (`/login`, `/auth/verify`,
  `/apply`, `/me/application/withdraw`,
  `/organizer/applications/:id/{approve,reject,revert}`,
  `/organizer/budget`); CSV at `/organizer/applications.csv`.
- ScholarshipOrganizerAuthMiddleware returns 401/403 for the API
  routes instead of redirecting to cfp.tryswift.jp like the existing
  OrganizerOnlyMiddleware does for SSR contexts.
- Magic-link flow: `POST /login` queues the link and 303-redirects to
  the page in `redirect_to`; the verify endpoint sets the
  `student_auth_token` cookie on `.tryswift.jp` and redirects back to
  the static portal so the cookie is visible on the next request.
- ScholarshipDTOs+Content adds Vapor `Content` conformance to the
  shared DTOs so the controller can encode them directly.

Cleanup:
- Delete StudentHostOnlyMiddleware, StudentLocaleMiddleware, and
  NotStudentHostMiddleware along with the three SSR controllers and
  ScholarshipRoutes — none of those make sense once the static site
  is the only thing serving student.tryswift.jp.
- HostRoutingMiddleware drops `isStudentHost`; only the sponsor
  portal still needs host detection.
- routes.swift mounts ScholarshipAPIController under
  `api.grouped("scholarship")` and removes the host-gated registration.

Tests:
- ScholarshipApplicationFlowTests now hits `/api/v1/scholarship/...`
  and adds a /me endpoint test (200 with cookie, 401 without).
- Full suite green: 137 tests in 27 suites.
